refactor: extract kolu-github integration package
Mirror the kolu-git pattern: gh-specific schemas + pure helpers
(classifyGhError, deriveCheckStatus, prResultEqual) move to a new
packages/integrations/github/ leaf. kolu-common depends on kolu-github
(same direction as the existing kolu-git dep); kolu-common/pr subpath
stays as a browser-light re-export of the neutral surface only.

Server's meta/github.ts shrinks from 346 to ~180 lines — keeps the
runtime wiring (KOLU_GH_BIN env load, `gh pr view` spawn, polling,
channel publisher) and imports the pure fns from kolu-github.

Neutral `PrResult` / `PrUnavailableSource` scaffolding lives in the
new package alongside the gh schemas, because PrResult.ok.value is
`GitHubPrInfo`-shaped today and splitting would invert the package
dep direction. Promote the neutrals out when bkt lands
(srid/agency#10).
